Add newSolverWithConfig
This adds `newSolverWithConfig`, which is intended to offer the maximal amount
of flexibility in configuring how to interact with the SMT solver subprocess.
All of the configuration-related options are now folded into `Config` and
`SolverLogger` data types, and the existing `newSolver`/`newSolverNotify`
functions are now defined in terms of `newSolverWithConfig`. See the Haddocks
for more information.

Fixes #26.